fix: more test data
