+++ data-generation/dois.js
@@ -1,0 +1,72 @@
+// SPDX-FileCopyrightText: 2023 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
+// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+//
+// SPDX-License-Identifier: Apache-2.0
+
+const conceptDois = [
+	'10.5281/zenodo.1034002',
+	'10.5281/zenodo.1043306',
+	'10.5281/zenodo.1043937',
+	'10.5281/zenodo.1045122',
+	'10.5281/zenodo.1045193',
+	'10.5281/zenodo.1051033',
+	'10.5281/zenodo.1051039',
+	'10.5281/zenodo.1051130',
+	'10.5281/zenodo.1083950',
+	'10.5281/zenodo.1145886',
+	'10.5281/zenodo.1149010',
+	'10.5281/zenodo.1162057',
+	'10.5281/zenodo.1404735',
+	'10.5281/zenodo.1435860',
+	'10.5281/zenodo.1436372',
+	'10.5281/zenodo.1436464',
+	'10.5281/zenodo.1462264',
+	'10.5281/zenodo.3234136',
+	'10.5281/zenodo.4050179',
+	'10.5281/zenodo.47756',
+	'10.5281/zenodo.594525',
+	'10.5281/zenodo.594559',
+	'10.5281/zenodo.594690',
+	'10.5281/zenodo.596839',
+	'10.5281/zenodo.597226',
+	'10.5281/zenodo.597238',
+	'10.5281/zenodo.597262',
+	'10.5281/zenodo.597793',
+	'10.5281/zenodo.597984',
+	'10.5281/zenodo.598013',
+	'10.5281/zenodo.598204',
+	'10.5281/zenodo.6379973',
+	'10.5281/zenodo.6532349',
+	'10.5281/zenodo.832894',
+	'10.5281/zenodo.909307',
+	'10.5281/zenodo.910447',
+	'10.5281/zenodo.910905',
+	'10.5281/zenodo.926819',
+	'10.5281/zenodo.997272',
+	'10.5281/zenodo.997332',
+];
+
+const dois = [
+	'10.1007/978-3-030-55874-1_30',
+	'10.1016/j.future.2018.08.004',
+	'10.1017/9781009085809',
+	'10.1017/S0033291718004038',
+	'10.1017/S1368980018001258',
+	'10.1029/2018MS001600',
+	'10.1109/eScience.2016.7870925',
+	'10.1145/2996913.2997005',
+	'10.1175/BAMS-D-19-0337.1',
+	'10.1186/s12966-019-0834-1',
+	'10.1515/itit-2019-0040',
+	'10.4233/uuid:4bb38399-9267-428f-b10a-80b86e101f23',
+	'10.5194/egusphere-egu21-4805',
+	'10.5194/ems2022-105',
+	'10.5194/esd-12-253-2021',
+	'10.5194/gmd-13-4205-2020',
+	'10.5281/zenodo.1149011',
+	'10.5281/zenodo.4075237',
+	'10.5281/zenodo.5748141',
+	'10.5334/dsj-2022-010',
+];
+
+export {conceptDois, dois};
