+++ frontend/components/software/__mocks__/softwareReleases.json.license
@@ -1,5 +1,7 @@
 SPDX-FileCopyrightText: 2022 - 2023 Dusan Mijatovic (dv4all)
 SPDX-FileCopyrightText: 2022 - 2023 dv4all
+SPDX-FileCopyrightText: 2023 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
+SPDX-FileCopyrightText: 2023 Netherlands eScience Center
 
 SPDX-License-Identifier: Apache-2.0
 SPDX-License-Identifier: CC-BY-4.0
